refactor(direct-control): extract map read types
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only map summary, plot snapshot, and map grid input/result type ownership from packages/civ7-direct-control/src/index.ts into packages/civ7-direct-control/src/play/map/types.ts while keeping facade type re-exports in index.ts.

Preserves public type contracts and does not change map read validation, hidden-info policy semantics, source strings, runtime behavior, visibility or GameInfo types, procedure schemas, telemetry, AI ingestion, hotseat runtime proof, CLI semantic projection, Effect/oRPC procedure-core work, or Task 2.9.4 matrix-row acceptance.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- map-and-visibility.test.ts public-api.test.ts setup-and-lifecycle.test.ts
- bun run --cwd packages/cli test -- game.control.test.ts
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- ownership/no-cycle rg scans

Local package/type relocation proof only; no runtime/live-game, hotseat product-path, AI-ingestion, semantic CLI, telemetry, or procedure-core proof claimed.
