+++ openspec/changes/civ7-support-direct-control-modularization/tasks.md
@@ -382,6 +382,14 @@ runtime/direct-control claims.
         types, procedure schemas, operation/proof telemetry, hotseat runtime
         proof, AI ingestion, CLI semantic projection, Effect/oRPC
         procedure-core work, and Task 2.9.4 matrix-row acceptance pending.
+  - [x] 4.8.12 Extract map read type owner while keeping public facade type
+        re-exports in `index.ts`, preserving map summary, plot snapshot, map
+        grid input/result, and full-map-grid chunk/input/result public
+        contracts while leaving visibility, GameInfo, summary, setup,
+        tactical, operation, and ready result types, procedure schemas,
+        operation/proof telemetry, hotseat runtime proof, AI ingestion, CLI
+        semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4
+        matrix-row acceptance pending.
 - [ ] 4.9 Import or explicitly cite oRPC architecture authority before any
       oRPC implementation; current support branch has no tracked
       `.agents/skills/civ7-orpc-control-architecture` or
